Added NPC Shooting and Prefabs
Updated the NPC script to allow for enemies to shoot at the player, will need to be updated in the future once we know how the player script takes damage. Along this, there are now prefabs for enemies that shoot, stronger enemies, and NPCs that ignore the player unless attacked.